moved eventnotify from init
